Set run frequency to period supported by github's garbage cron cpu time allowances
Github's idiocy of only allowing at minimum 5 minutes between a run, in which case, why even call this a cron job? To call this a cron job configuration and to use the kind of syntax for it is misleading, and deceptive, given that cron tax syntax allow for frequencies far smaller than the limit stood by Github clowns.